implementing border with different colors and radius

+++ core/src/base_types.hpp
@@ -7,10 +7,12 @@ struct rel {
   rel(float val): val(val) {};
   float val;
 };
+
 struct abs {
-  abs(float val): val(val) {};
-  float val;
+  abs(int val): val(val) {};
+  int val;
 };
+
 struct none {};
 
 }  // namespace value
@@ -21,14 +23,14 @@ struct Scale {
 };
 
 struct Size {
-	float width = 0;
-	float height = 0;
+	int width = 0;
+	int height = 0;
 	static bool is_equal(Size a, Size b);
 };
 
 struct Position {
-	float left = 0;
-	float top = 0;
+	int left = 0;
+	int top = 0;
 	static Position add(Position a, Position b);
 };
